+++ tests/ounit_tests/ounit_tests_main.ml
@@ -31,6 +31,7 @@ let suites =
       Ounit_js_analyzer_tests.suites;
       Ounit_flow_parser_tests.suites;
       Ounit_jsx_loc_tests.suites;
+      Ounit_analysis_find_files_tests.suites;
       Ounit_analysis_config_tests.suites;
       Ounit_analysis_references_tests.suites;
       Ounit_ffi_inclusion_tests.suites;